Extract transaction logic into separate file
